+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.4] - 2021-03-15
+
+### Added
+
+#### Bonus 1 - Save the datas
+
+- Create profile entity
+- EN translations
+- Make sure that this data is persistent (by id parameter in query)
+
 ## [0.0.3] - 2021-03-15
 
 ### Added